QDMA Linux reference driver for v2023.1.2 release

+++ QDMA/linux-kernel/driver/libqdma/libqdma_export.h
@@ -404,7 +404,7 @@ struct qdma_dev_conf {
 	/** pointer to pci_dev */
 	struct pci_dev *pdev;
 	/** Maximum number of queue pairs per device */
-	unsigned short qsets_max;
+	u32 qsets_max;
 	/** Reserved */
 	unsigned short rsvd2;
 	/**
@@ -1345,6 +1345,8 @@ int qdma_queue_get_config(unsigned long dev_hndl, unsigned long id,
  * Display all configured queues in a string buffer
  *
  * @param dev_hndl	dev_hndl returned from qdma_device_open()
+ * @param qidx		Queue index
+ * @param num_q		Number of Queues to list
  * @param buflen	length of the input buffer
  * @param buf		message buffer
  *
